app/requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
psycopg2-binary==2.9.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-redis-lock==4.0.0
pytz==2025.2
redis==6.2.0
requests==2.32.4
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
